webapp/requirements.txt
matplotlib
numpy
opencv-python
pillow
scikit-image
scikit-learn
torch
torchvision
gradio
onnxruntime
rasterio